feat: [prd-embedded-observer-graph-editor-006] - [Add regression coverage for the embedded editing path]
