+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -73,7 +73,7 @@ work inputs.
 - Shared workflow-activity edit handles should be layered onto the observer graph node data in `ui/src/features/workflow-activity/lib/react-flow-current-activity-card-graph.ts` and `...editor-handles.ts`, while unsupported rendered edge families keep invisible fallback React Flow handles in `current-activity-node-shell.tsx`; this keeps workstation/work-state editing on one authoritative anchor-id scheme without breaking existing observer-only edges such as resource lanes.
 - Shared observer-mode route creation should synthesize supported draft route additions and removals from `draft.edgeChanges` back onto the existing observer `GraphLayout` in `react-flow-current-activity-card-graph.ts`, then let `react-flow-current-activity-card-edges.ts` style those synthesized edges as pending additions; do not swap back to the retired editor-only graph just to preview unsaved workstation/work-state routes.
 - Embedded observer-graph editor regressions should stay split across three seams: keep anchor-id plus pending-edge projection and styling assertions in `ui/src/features/workflow-activity/lib/react-flow-current-activity-card-graph.test.ts`, keep draft-mutation and invalid-connection coverage in `ui/src/features/workflow-activity/hooks/react-flow-current-activity-card-editor-controllers.test.tsx`, and use `ui/src/features/workflow-activity/components/react-flow-current-activity-card.test.tsx` for the real DOM proof that shared-surface handle buttons and pending draft state stay visible without React Flow handle-attachment errors.
-- Replacement-topology layout-failure fallback coverage for the shared observer editor should live in a lightweight hook seam such as `ui/src/features/workflow-activity/components/react-flow-current-activity-card-view-model.test.tsx`, not in the heavier `react-flow-current-activity-card.coverage.test.tsx` file where duplicate async rerender cases widen fork-worker teardown risk.
+- Replacement-topology layout-failure fallback coverage for the shared observer editor should live in the dedicated graph-layout hook seam `ui/src/features/workflow-activity/hooks/react-flow-current-activity-card-graph-layout.test.tsx`, not in the full `react-flow-current-activity-card.tsx` import path or the heavier `react-flow-current-activity-card.coverage.test.tsx` file where duplicate async rerender cases widen fork-worker teardown risk.
 - Shared observer-editor DOM assertions should expect repeated accessible handle labels across multiple workstation nodes and use plural role queries or scoped counts rather than singleton `getByRole(...)`; exact pending-edge class and stroke assertions are more stable in the graph seam than in the heavy card DOM suite, where the card-level proof should focus on customer-visible dirty state plus the absence of React Flow handle errors.
 - Tests that mount `ui/src/features/workflow-activity/components/workflow-activity-bento-card.tsx`, App-level dashboard shells, or similar current-activity graph shells should mirror `components/react-flow-current-activity-card.test.tsx` and stub `useCurrentFactoryDocument`, `useSaveCurrentFactory`, and `useFactoryGraphDraftState`; the graph-editor hook stack now mounts in observe mode as part of the shared chrome.
 - Workflow-activity tests that render localized shells such as `workflow-activity-bento-card.tsx` should derive expected accessible names from the same message catalog helpers the component uses, rather than hardcoding English viewport or toolbar labels when the test passes a non-default `locale`.
