+++ ui/src/features/workflow-activity/hooks/react-flow-current-activity-card-graph-layout.test.tsx
@@ -1,14 +1,11 @@
 import { act, renderHook, waitFor } from "@testing-library/react";
 
-import {
-  semanticWorkflowDashboardSnapshot,
-  singleNodeDashboardSnapshot,
-} from "../../../components/dashboard/test-fixtures";
+import { singleNodeDashboardSnapshot } from "../../../components/dashboard/test-fixtures";
 import type { GraphLayout } from "../../flowchart/lib/layout";
-import { useCurrentActivityGraphViewModel } from "./react-flow-current-activity-card";
+import { useCurrentActivityGraphLayout } from "./react-flow-current-activity-card-graph-layout";
 
 type BuildGraphLayout = (
-  topology: typeof semanticWorkflowDashboardSnapshot.topology,
+  topology: typeof singleNodeDashboardSnapshot.topology,
 ) => Promise<GraphLayout>;
 
 const { actualBuildGraphLayoutRef, mockBuildGraphLayout } = vi.hoisted(() => ({
@@ -33,40 +30,7 @@ vi.mock("../../flowchart/lib/layout", async () => {
   };
 });
 
-function createEditorStub(overrides: Record<string, unknown> = {}) {
-  return {
-    activeTool: null,
-    canInteractWithEditor: false,
-    draftState: {
-      draft: {
-        additions: {
-          resources: [],
-          workers: [],
-          workStates: [],
-          workTypes: [],
-          workstations: [],
-        },
-        edgeChanges: {
-          additions: [],
-          removals: [],
-        },
-        removals: {
-          resources: [],
-          workers: [],
-          workStates: [],
-          workTypes: [],
-          workstations: [],
-        },
-      },
-    },
-    editorMode: false,
-    handleConnectionAnchorClick: vi.fn(),
-    pendingConnectionSource: null,
-    ...overrides,
-  };
-}
-
-describe("useCurrentActivityGraphViewModel", () => {
+describe("useCurrentActivityGraphLayout", () => {
   beforeEach(() => {
     mockBuildGraphLayout.mockReset();
     window.localStorage.clear();
@@ -74,10 +38,10 @@ describe("useCurrentActivityGraphViewModel", () => {
 
   it("falls back to the empty graph outcome when a replacement current-activity layout fails", async () => {
     const loadedSnapshot = structuredClone(singleNodeDashboardSnapshot);
-    const rejectedSnapshot = structuredClone(semanticWorkflowDashboardSnapshot);
-    const onSelectStateNode = vi.fn();
-    const onSelectWorkID = vi.fn();
-    const onSelectWorkstation = vi.fn();
+    const rejectedSnapshot = structuredClone(singleNodeDashboardSnapshot);
+    rejectedSnapshot.topology.workstation_nodes_by_id[
+      rejectedSnapshot.topology.workstation_node_ids[0]
+    ].workstation_name = "Rejected layout workstation";
 
     mockBuildGraphLayout.mockImplementation(async (topology) => {
       if (topology === rejectedSnapshot.topology) {
@@ -92,16 +56,7 @@ describe("useCurrentActivityGraphViewModel", () => {
     });
 
     const { result, rerender } = renderHook(
-      ({ snapshot }) =>
-        useCurrentActivityGraphViewModel({
-          editor: createEditorStub() as never,
-          now: Date.parse("2026-04-08T12:00:00Z"),
-          onSelectStateNode,
-          onSelectWorkID,
-          onSelectWorkstation,
-          selection: null,
-          snapshot,
-        }),
+      ({ snapshot }) => useCurrentActivityGraphLayout(snapshot),
       {
         initialProps: {
           snapshot: loadedSnapshot,
